requirements.txt
astroid==3.2.2
colorama==0.4.6
dill==0.3.8
isort==5.13.2
mccabe==0.7.0
pillow==10.3.0
platformdirs==4.2.2
pylint==3.2.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.11.0
